Fixed simulator warning
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@3743 a1c6a512-1295-4272-9138-f99709370657

+++ apps/settings.c
@@ -1679,7 +1679,7 @@ bool set_option(char* string, void* variable, enum optiontype type,
             case BUTTON_MENU:
 #endif
                 if (((type==INT) && (*intvar != oldval)) ||
-                    ((type==BOOL) && (*boolvar != oldval))) {
+                    ((type==BOOL) && ((int)*boolvar != oldval))) {
                     if (type==INT)
                         *intvar=oldval;
                     else
